refactor: slack refactor
