--- slack-service/src/main/java/com/gaebal_easy/client/slack/exception/SlackException.java
@@ -1,4 +1,0 @@
-package com.gaebal_easy.client.slack.exception;
-
-public class SlackException {
-}
