+++ slack-service/src/main/java/com/gaebal_easy/client/slack/presentation/SlackMessageController.java
@@ -9,7 +9,6 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.gaebal_easy.client.slack.application.service.GeminiService;
 import com.gaebal_easy.client.slack.application.service.SlackMessageService;
 import com.gaebal_easy.client.slack.presentation.dto.SlackResponse;
 
@@ -22,12 +21,7 @@ import lombok.RequiredArgsConstructor;
 public class SlackMessageController {
 
 	private final SlackMessageService slackMessageService;
-	private final GeminiService geminiService;
-	//
-	// @PostMapping
-	// public void sendSlack () {
-	// 	slackService.sendMessage("hi", "U08J95TTQA0");
-	// }
+
 	@GetMapping("/getSlack")
 	@PreAuthorize("hasRole('MASTER')")
 	public ResponseEntity<ApiResponseData<SlackResponse.GetSlackMessagesResponse>> getSlackMessages (
