+++ slack-service/src/main/java/com/gaebal_easy/client/slack/presentation/kafka/SlackMessageListener.java
@@ -1,6 +1,4 @@
-package com.gaebal_easy.client.slack.presentation;
-
-import java.util.UUID;
+package com.gaebal_easy.client.slack.presentation.kafka;
 
 import org.springframework.kafka.annotation.KafkaListener;
 import org.springframework.stereotype.Service;
@@ -9,7 +7,6 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.gaebal_easy.client.slack.application.service.GeminiService;
 import com.gaebal_easy.client.slack.presentation.dto.SlackMessageInfoDTO;
-import com.gaebal_easy.client.slack.presentation.dto.SlackRequest;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
